+++ core/src/main/java/io/questdb/griffin/engine/functions/groupby/AvgDecimal128Rescale256GroupByFunction.java
@@ -106,7 +106,7 @@ class AvgDecimal128Rescale256GroupByFunction extends Decimal256Function implemen
     @Override
     public void getDecimal128(Record rec, Decimal128 sink) {
         if (calc(rec)) {
-            sink.ofRaw((short) decimal256A.getLh(), decimal256A.getLl());
+            sink.ofRaw(decimal256A.getLh(), decimal256A.getLl());
         } else {
             sink.ofRawNull();
         }
